+++ src/main/java/com/hyoguoo/paymentplatform/core/common/metrics/PaymentStatusGaugeMetrics.java
@@ -1,4 +1,4 @@
-package com.hyoguoo.paymentplatform.payment.application.metrics;
+package com.hyoguoo.paymentplatform.core.common.metrics;
 
 import com.hyoguoo.paymentplatform.core.common.service.port.LocalDateTimeProvider;
 import com.hyoguoo.paymentplatform.payment.application.port.PaymentEventRepository;
