+++ src/main/java/com/hyoguoo/paymentplatform/payment/scheduler/PaymentScheduler.java
@@ -1,6 +1,6 @@
 package com.hyoguoo.paymentplatform.payment.scheduler;
 
-import com.hyoguoo.paymentplatform.payment.application.metrics.SchedulerMetrics;
+import com.hyoguoo.paymentplatform.core.common.metrics.SchedulerMetrics;
 import com.hyoguoo.paymentplatform.payment.scheduler.port.PaymentExpirationService;
 import com.hyoguoo.paymentplatform.payment.scheduler.port.PaymentRecoverService;
 import lombok.RequiredArgsConstructor;
